Moving Beacons things to 'beacons' dir
