+++ src/test/java/uk/gov/mca/beacons/service/registrations/BeaconServiceTest.java
@@ -16,6 +16,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.mca.beacons.service.beacons.BeaconsRelationshipMapper;
+import uk.gov.mca.beacons.service.beacons.BeaconsService;
 import uk.gov.mca.beacons.service.dto.BeaconDTO;
 import uk.gov.mca.beacons.service.model.Activity;
 import uk.gov.mca.beacons.service.model.Beacon;
